+++ FridgeExercise/Shelf.cs
@@ -14,11 +14,11 @@ namespace FridgeExercise
 		public int ShelfID { get; set; }
 		private int CounterID = 1;
 		public int ShelfsFloor { get; set; }
-		public int SpaceInShelf { get; set; }
+		public double SpaceInShelf { get; set; }
 		public List<Item> Items { get; set; }
-		public int CurrentSpaceInShelf { get; set; }
+		public double CurrentSpaceInShelf { get; set; }
 
-		public Shelf(int shelfsFloor,int spaceInShelf/*, List<Item> items*/) {
+		public Shelf(int shelfsFloor, double spaceInShelf/*, List<Item> items*/) {
 			ShelfID = CounterID;
 			CounterID++;
 			ShelfsFloor = shelfsFloor;
